+++ app/routes.py
@@ -463,88 +463,6 @@ def current_member_bids():
             return jsonify({'message': 'Denied bid not found for this member with the provided bid ID'}), 404
 
 
-@app.route('/api/vehicle-purchase/new-vehicle-no-finance/bid-accepted', methods=['POST'])
-# this endpoint is used for the customer to Purchase their cars AFTER their bid had been accepted
-# here we only deal with the Purchases and Payments Table
-def vehicle_purchase_bid_accepted():
-    try:
-        # make sure the member is logged in and has an account to purchase a vehicle and it logged in
-        member_session_id = session.get('member_session_id')
-        if member_session_id is None:
-            return jsonify({'message': 'Invalid session'}), 400
-
-        # get information sent from the frontend to here
-        # Frontend needs to send: bid_id, payment_option, member_id, payment_amount
-        data = request.json
-        bid_id = data.get('bid_id')
-        payment_option = data.get('payment_option')  # 'Card' or 'Check'
-        member_id = data.get('member_id')
-        payment_amount = data.get('payment_amount')
-
-        # returns the row in BID table that matches with our bid_id
-        bid = Bids.query.get(bid_id)
-        if not bid or bid.bidStatus != 'Confirmed':
-            return jsonify({'message': 'Bid not found or is Not Confirmed, Cannot continue with Purchase'}), 404
-
-        # returns vehicle information for purchase for the vehicle to be bought
-        vehicle = CarInfo.query.filter_by(VIN_carID=bid.VIN_carID).first()
-        if not vehicle:
-            return jsonify({'message': 'Vehicle not found for this bid'}), 404
-
-        # Validate and retrieve card or check information based on payment_option
-        if payment_option == 'Card':
-            card_number = data.get('card_number')
-            cvv = data.get('CVV')
-            expiration_date = data.get('expirationDate')
-            routing_number = None
-            account_number = None
-            regex_card_check(card_number, cvv, expiration_date)
-            if payment_amount > 5000:  # cards cannot pay more than 5000
-                return jsonify({'message': 'Choose A lower value for Bank Cards'}), 400
-        elif payment_option == 'Check':
-            routing_number = data.get('routing_number')
-            account_number = data.get('account_number')
-            card_number = None
-            cvv = None
-            expiration_date = None
-            regex_bank_acct_check(routing_number, account_number)
-        else:
-            return jsonify({'message': 'Invalid payment option.'}), 400
-
-        # Retrieve vehicle cost
-        # vehicle_cost = return_vehicle_cost(vehicle_vin) # no need because the cost is based on the confirmed bid
-        total_valuePaid = bid.bidValue - payment_amount
-        # signature = get_signature()
-
-        # create a payment entry to insert into the DB for the new payment/purchase information
-        new_payment = Payments(
-            paymentStatus='Confirmed',
-            valuePaid=payment_amount,
-            valueToPay=total_valuePaid,
-            initialPurchase=datetime.now(),
-            lastPayment=datetime.now(),
-            paymentType=payment_option,
-            cardNumber=card_number,
-            expirationDate=expiration_date,
-            CVV=cvv,
-            routingNumber=routing_number,
-            bankAcctNumber=account_number,
-            memberID=member_id
-        )
-        db.session.add(new_payment)
-        db.session.commit()
-
-        new_purchase = Purchases(
-            VIN_carID=vehicle.VIN_carID,
-            memberID=member_id,
-            confirmationNumber=confirmation_number_generation(),  # You may generate a confirmation number here
-            # signature=signature_val
-        )
-        db.session.add(new_purchase)
-        db.session.commit()
-        return jsonify({'message': 'Vehicle purchase processed successfully.'}), 200
-    except Exception as e:
-        return jsonify({'error': 'Bid not found for the specified member and vehicle, could not purchase vehicle'}), 404
 
 
 @app.route('/api/manager/customer-payment-report', methods=['GET'])
@@ -775,61 +693,20 @@ def update_ssn(member_session_id):
     return True
 
 
-@app.route('/api/vehicle-purchase/new-vehicle-no-finance', methods=['POST'])
-# POST request is used by the customer to purchase a new Vehicle at BID or MSRP with NO FINANCING
-def purchase_vehicle():
-    member_id = session.get('member_session_id')
-    if member_id is None:
-        return jsonify({'message': 'You need to log in to purchase a vehicle.'}), 403
-
+# not in use yet
+@app.route('/api/vehicle-purchase/signature', methods=['POST'])
+def get_signature():
     data = request.json
-    vehicle_vin = data.get('vehicle_vin')
-    if vehicle_vin is None:
-        return jsonify({'message': "Vehicle VIN is missing."}), 400
-
-    # card_number = None
-    # cvv = None
-    # expiration_date = None
-
-    payment_method = data.get('payment_method')
-    payment_amount = data.get('payment_amount')
-    payment_option = "Check"
-
-    try:
-        vehicle_cost = return_vehicle_cost(vehicle_vin)
-    except ValueError as ve:
-        return jsonify({'error': str(ve)}), 404  # Vehicle not found
-    
-    if payment_method == 'MSRP':
-        routing_number = data.get('routing_number')
-        account_number = data.get('account_number')
-
-        if check_ssn() is False:
-            return jsonify({'message': 'SSN number is in Invalid Format.'}), 401
-
-        if regex_bank_acct_check(routing_number, account_number) is False:
-            return jsonify({'message': 'Routing Number and/or Account Number are in Invalid Formats.'}), 401
-
-        try:
-            return msrp_vehicle_purchase_no_financing(vehicle_vin, payment_amount, member_id, payment_option, vehicle_cost, routing_number, account_number)
-        except Exception as e:
-            return jsonify({'error': f'Error processing MSRP payment: {str(e)}'}), 500
-
+    signature = int(data.get('signature'))  # yes = 1, no = 0
+    if signature == 0:
+        return 'No'
+    elif signature == 1:
+        return 'Yes'
     else:
-        bid_value = data.get('bidValue')
-        bid_status = 'Processing'
-        try:
-            return bid_insert_no_financing(member_id, bid_value, bid_status)
-        except Exception as e:
-            return jsonify({'error': f'Error processing bid: {str(e)}'}), 500
+        return jsonify({'message': 'Invalid VALUE'}), 400
 
 
-# --- left off here ---
-@app.route('/api/vehicle-purchase/cart', methods=['POST'])
-# i have no clue how to implement this, i was thinking about this one being implemented based on indiv. sessions
-# and not needing another DB table for member carts
-def purchase_cart():
-    ...
+
 
 
 def regex_card_check(card_number: str, cvv: str, expiration_date: str) -> bool:
@@ -866,65 +743,6 @@ def regex_ssn(ssn: str) -> bool:
         return False
     return True
 
-
-def msrp_vehicle_purchase_no_financing(vehicle_vin, payment_amount, member_id, payment_option, vehicle_cost, routing_number, account_number):
-    try:
-        # Validate signature
-        signature_val = get_signature()
-        if signature_val not in ['Yes', 'No']:
-            return jsonify({'error': 'Invalid signature value.'}), 400
-        elif signature_val == 'No':
-            return jsonify({'error': 'You must provide a signature to purchase a vehicle.'}), 401
-
-        # Calculate payment values
-        valuePaid_value = payment_amount
-        valueToPay_value = vehicle_cost - payment_amount
-
-        # Add payment record
-        new_payment = Payments(
-            paymentStatus='Confirmed',
-            valuePaid=valuePaid_value,
-            valueToPay=valueToPay_value,
-            initialPurchase=datetime.now(),
-            lastPayment=datetime.now(),
-            paymentType=payment_option,
-            servicePurchased='Vehicle Purchase',
-            # cardNumber=card_number,
-            # expirationDate=expiration_date,
-            # CVV=cvv,
-            routingNumber=routing_number,
-            bankAcctNumber=account_number,
-            memberID=member_id,
-            financingID=None
-        )
-        db.session.add(new_payment)
-
-        # Commit payment record
-        db.session.commit()
-
-        # Add purchase record
-        new_purchase = Purchases(
-            VIN_carID=vehicle_vin,
-            memberID=member_id,
-            confirmationNumber=confirmation_number_generation(),
-            # purchaseType='Vehicle/Add-on Purchase',
-            # purchaseDate=datetime.now(),
-            # signature=signature_val
-        )
-        db.session.add(new_purchase)
-
-        # Commit purchase record
-        db.session.commit()
-
-        return jsonify({'message': 'Vehicle purchase processed successfully.'}), 200
-
-    except IntegrityError:
-        db.session.rollback()
-        return jsonify({'error': 'Integrity error occurred. Please check your input data.'}), 400
-
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'error': f'Error processing vehicle purchase: {str(e)}'}), 500
 
 
 def return_vehicle_cost(vehicle_vin):
@@ -1011,312 +829,456 @@ def adjust_loan_with_downpayment(vehicle_cost, down_payment):
     return loan_amount
 
 
-@app.route('/api/vehicle-purchase/new-vehicle-purchase-finance/re-evaluate', methods=['POST'])
-def reevaluate_finance():
-    # used as a means of re evaluating the loan and whether the user wants to or not and it will lead to a new downpayment
-    data = request.json
-    reevaluate_loan = data.get('reevaluate_loan')  # no == 0, yes == 1
-    return str(reevaluate_loan)
+# @app.route('/api/vehicle-purchase/new-vehicle-purchase-finance/re-evaluate', methods=['POST'])
+# def reevaluate_finance():
+#     # used as a means of re evaluating the loan and whether the user wants to or not and it will lead to a new downpayment
+#     data = request.json
+#     reevaluate_loan = data.get('reevaluate_loan')  # no == 0, yes == 1
+#     return str(reevaluate_loan)
 
 
-''' API to use to purchase a vehicle at MSRP with financing from the dealership'''
+# ''' API to use to purchase a vehicle at MSRP with financing from the dealership'''
 
 
-@app.route('/api/vehicle-purchase/new-vehicle-purchase-finance', methods=['POST'])
-def new_vehicle_purchase_finance():
-    # here we deal with Financing, Purchases and Payments table
-    # HERE we deal with purchases of vehicles WITH FINANCING
-    try:
-        # customer auth for making sure they are logged in and have an account
-        member_session_id = session.get('member_session_id')
-        if member_session_id is None:
-            return jsonify({'message': 'Invalid session'}), 400
+# @app.route('/api/vehicle-purchase/new-vehicle-purchase-finance', methods=['POST'])
+# def new_vehicle_purchase_finance():
+#     # here we deal with Financing, Purchases and Payments table
+#     # HERE we deal with purchases of vehicles WITH FINANCING
+#     try:
+#         # customer auth for making sure they are logged in and have an account
+#         member_session_id = session.get('member_session_id')
+#         if member_session_id is None:
+#             return jsonify({'message': 'Invalid session'}), 400
 
-        # frontend needs to send these values to the backend
-        data = request.json
-        vehicle_vin = data.get('vehicle_vin')
-        payment_method = data.get('payment_method')
-        member_id = data.get('member_id')
-        down_payment = data.get('down_payment')
-        monthly_income = data.get('monthly_income')
+#         # frontend needs to send these values to the backend
+#         data = request.json
+#         vehicle_vin = data.get('vehicle_vin')
+#         payment_method = data.get('payment_method')
+#         member_id = data.get('member_id')
+#         down_payment = data.get('down_payment')
+#         monthly_income = data.get('monthly_income')
 
-        if payment_method == 'CARD':
-            card_number = data.get('card_number')
-            cvv = data.get('cvv')
-            expiration_date = data.get('expirationDate')
-            routingNumber = None
-            bankAcctNumber = None
-            if down_payment > 5000:
-                return jsonify({'message': 'Card payments are limited to $5000. The rest must be paid in person at '
-                                           'the dealership.'}), 400
-        else:
-            routingNumber = data.get('routingNumber')
-            bankAcctNumber = data.get('bankAcctNumber')
-            card_number = None
-            cvv = None
-            expiration_date = None
+#         if payment_method == 'CARD':
+#             card_number = data.get('card_number')
+#             cvv = data.get('cvv')
+#             expiration_date = data.get('expirationDate')
+#             routingNumber = None
+#             bankAcctNumber = None
+#             if down_payment > 5000:
+#                 return jsonify({'message': 'Card payments are limited to $5000. The rest must be paid in person at '
+#                                            'the dealership.'}), 400
+#         else:
+#             routingNumber = data.get('routingNumber')
+#             bankAcctNumber = data.get('bankAcctNumber')
+#             card_number = None
+#             cvv = None
+#             expiration_date = None
 
-        credit_score = creditScoreGenerator()
-        vehicle_cost = return_vehicle_cost(vehicle_vin)
-        if vehicle_cost == -1:
-            return jsonify({'message': 'Vehicle Not Listed'}), 400
+#         credit_score = creditScoreGenerator()
+#         vehicle_cost = return_vehicle_cost(vehicle_vin)
+#         if vehicle_cost == -1:
+#             return jsonify({'message': 'Vehicle Not Listed'}), 400
 
-        total_cost = adjust_loan_with_downpayment(vehicle_cost, down_payment)
-        financing_loan_amount = financingValue(total_cost, monthly_income, credit_score)
+#         total_cost = adjust_loan_with_downpayment(vehicle_cost, down_payment)
+#         financing_loan_amount = financingValue(total_cost, monthly_income, credit_score)
 
-        # Loan eligibility
-        loan_eligibility = check_loan_eligibility(financing_loan_amount, monthly_income)
-        if not loan_eligibility:
-            # we want to check if the user wants to re-evaluate their loan through a new down payment amount
-            reevaluate_loan = int(reevaluate_finance())
-            if reevaluate_loan == 0:  # they cannot purchase the vehicle
-                return jsonify({'message': 'Your yearly income is not sufficient to take on this loan.'}), 400
-            elif reevaluate_loan == 1:  # they can and the frontend needs to send a new downpayment, this may need a new endpoint so LMK frontend
-                new_down_payment = data.get('new_down_payment')
-                total_cost = adjust_loan_with_downpayment(vehicle_cost, new_down_payment)
-                financing_loan_amount = financingValue(total_cost, monthly_income, credit_score)
-                loan_eligibility = check_loan_eligibility(financing_loan_amount, monthly_income)
-                # if true, we can continue to storing everything and all the values !!.
-                if not loan_eligibility:
-                    return jsonify({
-                        'message': 'Your yearly income is still not sufficient to take on this loan.'}), 400  # they cannot purchase this vehicle
-            else:
-                return jsonify({'message': 'Invalid Value'}), 400
+#         # Loan eligibility
+#         loan_eligibility = check_loan_eligibility(financing_loan_amount, monthly_income)
+#         if not loan_eligibility:
+#             # we want to check if the user wants to re-evaluate their loan through a new down payment amount
+#             reevaluate_loan = int(reevaluate_finance())
+#             if reevaluate_loan == 0:  # they cannot purchase the vehicle
+#                 return jsonify({'message': 'Your yearly income is not sufficient to take on this loan.'}), 400
+#             elif reevaluate_loan == 1:  # they can and the frontend needs to send a new downpayment, this may need a new endpoint so LMK frontend
+#                 new_down_payment = data.get('new_down_payment')
+#                 total_cost = adjust_loan_with_downpayment(vehicle_cost, new_down_payment)
+#                 financing_loan_amount = financingValue(total_cost, monthly_income, credit_score)
+#                 loan_eligibility = check_loan_eligibility(financing_loan_amount, monthly_income)
+#                 # if true, we can continue to storing everything and all the values !!.
+#                 if not loan_eligibility:
+#                     return jsonify({
+#                         'message': 'Your yearly income is still not sufficient to take on this loan.'}), 400  # they cannot purchase this vehicle
+#             else:
+#                 return jsonify({'message': 'Invalid Value'}), 400
 
-        # signature retrival value | ENUM just to ensure we go something and not left blank. | dont worry about this now, we are not that far into development
-        # signature = get_signature()
-        # if signature != 1:
-        #     return jsonify({'message': 'Please Insert Signature Value'})
+#         # signature retrival value | ENUM just to ensure we go something and not left blank. | dont worry about this now, we are not that far into development
+#         # signature = get_signature()
+#         # if signature != 1:
+#         #     return jsonify({'message': 'Please Insert Signature Value'})
 
-        # ik its long here but im tight on time imma just make it here and MAY refactor later
+#         # ik its long here but im tight on time imma just make it here and MAY refactor later
 
-        downPayment_value = total_cost - financing_loan_amount
-        valueToPay_value = total_cost - downPayment_value
-        paymentPerMonth_value = financing_loan_amount / 12
+#         downPayment_value = total_cost - financing_loan_amount
+#         valueToPay_value = total_cost - downPayment_value
+#         paymentPerMonth_value = financing_loan_amount / 12
 
-        # insertion of new Financing, Payments and Purchases Tables
-        new_financing = Financing(
-            memberID=member_id,
-            income=int(monthly_income) * 12,
-            credit_score=credit_score,
-            loan_total=financing_loan_amount,
-            down_payment=downPayment_value,
-            percentage=interest_rate(credit_score),
-            monthly_sum=paymentPerMonth_value,
-            remaining_months=48
-        )
-        db.session.add(new_financing)
-        db.session.commit()
+#         # insertion of new Financing, Payments and Purchases Tables
+#         new_financing = Financing(
+#             memberID=member_id,
+#             income=int(monthly_income) * 12,
+#             credit_score=credit_score,
+#             loan_total=financing_loan_amount,
+#             down_payment=downPayment_value,
+#             percentage=interest_rate(credit_score),
+#             monthly_sum=paymentPerMonth_value,
+#             remaining_months=48
+#         )
+#         db.session.add(new_financing)
+#         db.session.commit()
 
-        # DB insert for new purchase with financing
-        new_payment = Payments(
-            paymentStatus='Confirmed',
-            valuePaid=downPayment_value,
-            valueToPay=valueToPay_value,
-            initialPurchase=datetime.now(),
-            lastPayment=datetime.now(),
-            paymentType=payment_method,
-            servicePurchased='Vehicle Purchase',
-            cardNumber=card_number,
-            expirationDate=expiration_date,
-            CVV=cvv,
-            routingNumber=routingNumber,
-            bankAcctNumber=bankAcctNumber,
-            memberID=member_id
-        )
-        db.session.add(new_payment)
-        db.session.commit()
+#         # DB insert for new purchase with financing
+#         new_payment = Payments(
+#             paymentStatus='Confirmed',
+#             valuePaid=downPayment_value,
+#             valueToPay=valueToPay_value,
+#             initialPurchase=datetime.now(),
+#             lastPayment=datetime.now(),
+#             paymentType=payment_method,
+#             servicePurchased='Vehicle Purchase',
+#             cardNumber=card_number,
+#             expirationDate=expiration_date,
+#             CVV=cvv,
+#             routingNumber=routingNumber,
+#             bankAcctNumber=bankAcctNumber,
+#             memberID=member_id
+#         )
+#         db.session.add(new_payment)
+#         db.session.commit()
 
-        new_purchase = Purchases(
-            # no Bid ID since this is not a BID Operation
-            VIN_carID=vehicle_vin,
-            memberID=member_id,
-            confirmationNumber=confirmation_number_generation()  # You may generate a confirmation number here
-            # signature='YES'
-        )
-        db.session.add(new_purchase)
-        db.session.commit()
+#         new_purchase = Purchases(
+#             # no Bid ID since this is not a BID Operation
+#             VIN_carID=vehicle_vin,
+#             memberID=member_id,
+#             confirmationNumber=confirmation_number_generation()  # You may generate a confirmation number here
+#             # signature='YES'
+#         )
+#         db.session.add(new_purchase)
+#         db.session.commit()
 
-        # payment stub generation can occur through the means of functions above with endpoints
-        # /api/member
-        # /api/payments
+#         # payment stub generation can occur through the means of functions above with endpoints
+#         # /api/member
+#         # /api/payments
 
-        return jsonify({'message': 'Vehicle purchase with financing processed successfully.'}), 200
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'message': f'Error: {str(e)}'}), 500
-
-
-# not in use yet
-@app.route('/api/vehicle-purchase/signature', methods=['POST'])
-def get_signature():
-    data = request.json
-    signature = int(data.get('signature'))  # yes = 1, no = 0
-    if signature == 0:
-        return 'No'
-    elif signature == 1:
-        return 'Yes'
-    else:
-        return jsonify({'message': 'Invalid VALUE'}), 400
-
-
-@app.route('/api/vehicle-purchase/bid-confirmed-financed-purchase', methods=['POST'])
-def new_bid_purchase_finance():
-    # here we deal with the bid, purchases, payments and finance tables
-    # HERE we deal with Bid that are confimed and being purchased with FINANCING
-    try:
-        # customer auth for making sure they are logged in and have an account
-        member_session_id = session.get('member_session_id')
-        if member_session_id is None:
-            return jsonify({'message': 'Invalid session'}), 400
-
-        # frontend needs to send these values to the backend
-        data = request.json
-        bid_id = data.get('bid_id')
-        member_id = data.get('member_id')
-        payment_method = data.get('payment_method')
-        down_payment = data.get('down_payment')
-        monthly_income = data.get('monthly_income')
-
-        # returns data on the bid we need that the customer is purchasing
-        bid = Bids.query.get(bid_id)
-        if not bid or bid.bidStatus != 'Confirmed':
-            return jsonify({'message': 'Bid not found or is Not Confirmed, Cannot continue with Purchase'}), 404
-
-        # return purchasing information on the vehicle to be purchased, we need the information not COST since we are buying based on an approved BID
-        vehicle = CarInfo.query.filter_by(VIN_carID=bid.VIN_carID).first()
-        if not vehicle:
-            return jsonify({'message': 'Vehicle not found for this bid'}), 404
-
-        # more processing
-        if payment_method == 'CARD':
-            card_number = data.get('card_number')
-            cvv = data.get('cvv')
-            expiration_date = data.get('expirationDate')
-            routingNumber = None
-            bankAcctNumber = None
-            if down_payment > 5000:
-                return jsonify({
-                    'message': 'Card payments are limited to $5000. The rest must be paid in person at the dealership.'}), 400
-        else:
-            routingNumber = data.get('routingNumber')
-            bankAcctNumber = data.get('bankAcctNumber')
-            card_number = None
-            cvv = None
-            expiration_date = None
-
-        credit_score = creditScoreGenerator()  # Assuming this function exists and works correctly
-        vehicle_cost = vehicle.price
-        total_cost = adjust_loan_with_downpayment(vehicle_cost, down_payment)
-
-        # Loan Eligibility | the same as the other loan operations in buying the vehicle at MSRP with financing
-        financing_loan_amount = financingValue(total_cost, monthly_income, credit_score)
-        loan_eligibility = check_loan_eligibility(financing_loan_amount, monthly_income)
-        if not loan_eligibility:
-            reevaluate_loan = int(reevaluate_finance())
-            if reevaluate_loan == 0:
-                return jsonify({'message': 'Your yearly income is not sufficient to take on this loan.'}), 400
-            elif reevaluate_loan == 1:
-                new_down_payment = data.get('new_down_payment')
-                total_cost = adjust_loan_with_downpayment(vehicle_cost, new_down_payment)
-                financing_loan_amount = financingValue(total_cost, monthly_income, credit_score)
-                loan_eligibility = check_loan_eligibility(financing_loan_amount, monthly_income)
-                if not loan_eligibility:
-                    return jsonify({'message': 'Your yearly income is still not sufficient to take on this loan.'}), 400
-            else:
-                return jsonify({'message': 'Invalid Value'}), 400
-
-        # signature = get_signature()  # Assuming this function exists and works correctly
-        # if signature != 1:
-        #     return jsonify({'message': 'Please Insert Signature Value'}), 400
-
-        downPayment_value = total_cost - financing_loan_amount
-        valueToPay_value = total_cost - downPayment_value
-        paymentPerMonth_value = financing_loan_amount / 12
-
-        # inserts new values on the purchase based on Financing, Payments, Purchases
-        new_financing = Financing(
-            memberID=member_id,
-            income=monthly_income,
-            credit_score=credit_score,
-            loan_total=financing_loan_amount,
-            down_payment=downPayment_value,
-            percentage=interest_rate(credit_score),
-            monthly_sum=paymentPerMonth_value,
-            remaining_months=48
-        )
-        db.session.add(new_financing)
-        db.session.commit()
-
-        new_payment = Payments(
-            paymentStatus='Confirmed',
-            valuePaid=downPayment_value,
-            valueToPay=valueToPay_value,
-            initialPurchase=datetime.now(),
-            lastPayment=datetime.now(),
-            paymentType=payment_method,
-            cardNumber=card_number,
-            expirationDate=expiration_date,
-            CVV=cvv,
-            routingNumber=routingNumber,
-            bankAcctNumber=bankAcctNumber,
-            memberID=member_id,
-            financingID=new_financing.financingID
-        )
-        db.session.add(new_payment)
-        db.session.commit()
-
-        new_purchase = Purchases(
-            bidID=bid.bidID,
-            VIN_carID=vehicle.VIN_carID,
-            memberID=member_id,
-            confirmationNumber=confirmation_number_generation(),  # You may generate a confirmation number here
-            # signature=signature_val
-        )
-        db.session.add(new_purchase)
-        db.session.commit()
-
-        # no need to modify Bid Status since its already been confirmed.
-        return jsonify({'message': 'Vehicle purchase with financing processed successfully.'}), 200
-    except Exception as e:
-        db.session.rollback()
-        return jsonify({'message': f'Error: {str(e)}'}), 500
+#         return jsonify({'message': 'Vehicle purchase with financing processed successfully.'}), 200
+#     except Exception as e:
+#         db.session.rollback()
+#         return jsonify({'message': f'Error: {str(e)}'}), 500
 
 
 
-#Worked with /foward
-@app.route('/api/vehicle-purchase/new-bid-insert', methods=['POST'])
-# Adds a new bid to bid table
-def bid_insert_no_financing():
-    try:
-        # Extract data from the request
-        data = request.get_json()
-        required_fields = ['member_id', 'vin', 'bid_value']
-        
-        # Check if all required fields are present
-        missing_fields = [field for field in required_fields if field not in data]
-        if missing_fields:
-            return jsonify({'message': f'Error: Missing fields - {", ".join(missing_fields)}'}), 400
-        
-        # Extract data
-        member_id = data['member_id']
-        vin = data['vin']
-        bid_value = data['bid_value']
-        bid_status = 'Processing'
-        
-        # Create a new bid entry
-        new_bid = Bids(
-            memberID=member_id,
-            VIN_carID=vin,
-            bidValue=bid_value,
-            bidStatus=bid_status,
-            bidTimestamp=datetime.now()
-        )
 
-        db.session.add(new_bid)
-        db.session.commit()
-        return jsonify({'message': 'Bid successfully inserted.'}), 201
-    except Exception as e:
-        # Rollback the transaction in case of an error
-        db.session.rollback()
-        return jsonify({'message': f'Error: {str(e)}'}), 500
+# @app.route('/api/vehicle-purchase/bid-confirmed-financed-purchase', methods=['POST'])
+# def new_bid_purchase_finance():
+#     # here we deal with the bid, purchases, payments and finance tables
+#     # HERE we deal with Bid that are confimed and being purchased with FINANCING
+#     try:
+#         # customer auth for making sure they are logged in and have an account
+#         member_session_id = session.get('member_session_id')
+#         if member_session_id is None:
+#             return jsonify({'message': 'Invalid session'}), 400
+
+#         # frontend needs to send these values to the backend
+#         data = request.json
+#         bid_id = data.get('bid_id')
+#         member_id = data.get('member_id')
+#         payment_method = data.get('payment_method')
+#         down_payment = data.get('down_payment')
+#         monthly_income = data.get('monthly_income')
+
+#         # returns data on the bid we need that the customer is purchasing
+#         bid = Bids.query.get(bid_id)
+#         if not bid or bid.bidStatus != 'Confirmed':
+#             return jsonify({'message': 'Bid not found or is Not Confirmed, Cannot continue with Purchase'}), 404
+
+#         # return purchasing information on the vehicle to be purchased, we need the information not COST since we are buying based on an approved BID
+#         vehicle = CarInfo.query.filter_by(VIN_carID=bid.VIN_carID).first()
+#         if not vehicle:
+#             return jsonify({'message': 'Vehicle not found for this bid'}), 404
+
+#         # more processing
+#         if payment_method == 'CARD':
+#             card_number = data.get('card_number')
+#             cvv = data.get('cvv')
+#             expiration_date = data.get('expirationDate')
+#             routingNumber = None
+#             bankAcctNumber = None
+#             if down_payment > 5000:
+#                 return jsonify({
+#                     'message': 'Card payments are limited to $5000. The rest must be paid in person at the dealership.'}), 400
+#         else:
+#             routingNumber = data.get('routingNumber')
+#             bankAcctNumber = data.get('bankAcctNumber')
+#             card_number = None
+#             cvv = None
+#             expiration_date = None
+
+#         credit_score = creditScoreGenerator()  # Assuming this function exists and works correctly
+#         vehicle_cost = vehicle.price
+#         total_cost = adjust_loan_with_downpayment(vehicle_cost, down_payment)
+
+#         # Loan Eligibility | the same as the other loan operations in buying the vehicle at MSRP with financing
+#         financing_loan_amount = financingValue(total_cost, monthly_income, credit_score)
+#         loan_eligibility = check_loan_eligibility(financing_loan_amount, monthly_income)
+#         if not loan_eligibility:
+#             reevaluate_loan = int(reevaluate_finance())
+#             if reevaluate_loan == 0:
+#                 return jsonify({'message': 'Your yearly income is not sufficient to take on this loan.'}), 400
+#             elif reevaluate_loan == 1:
+#                 new_down_payment = data.get('new_down_payment')
+#                 total_cost = adjust_loan_with_downpayment(vehicle_cost, new_down_payment)
+#                 financing_loan_amount = financingValue(total_cost, monthly_income, credit_score)
+#                 loan_eligibility = check_loan_eligibility(financing_loan_amount, monthly_income)
+#                 if not loan_eligibility:
+#                     return jsonify({'message': 'Your yearly income is still not sufficient to take on this loan.'}), 400
+#             else:
+#                 return jsonify({'message': 'Invalid Value'}), 400
+
+#         # signature = get_signature()  # Assuming this function exists and works correctly
+#         # if signature != 1:
+#         #     return jsonify({'message': 'Please Insert Signature Value'}), 400
+
+#         downPayment_value = total_cost - financing_loan_amount
+#         valueToPay_value = total_cost - downPayment_value
+#         paymentPerMonth_value = financing_loan_amount / 12
+
+#         # inserts new values on the purchase based on Financing, Payments, Purchases
+#         new_financing = Financing(
+#             memberID=member_id,
+#             income=monthly_income,
+#             credit_score=credit_score,
+#             loan_total=financing_loan_amount,
+#             down_payment=downPayment_value,
+#             percentage=interest_rate(credit_score),
+#             monthly_sum=paymentPerMonth_value,
+#             remaining_months=48
+#         )
+#         db.session.add(new_financing)
+#         db.session.commit()
+
+#         new_payment = Payments(
+#             paymentStatus='Confirmed',
+#             valuePaid=downPayment_value,
+#             valueToPay=valueToPay_value,
+#             initialPurchase=datetime.now(),
+#             lastPayment=datetime.now(),
+#             paymentType=payment_method,
+#             cardNumber=card_number,
+#             expirationDate=expiration_date,
+#             CVV=cvv,
+#             routingNumber=routingNumber,
+#             bankAcctNumber=bankAcctNumber,
+#             memberID=member_id,
+#             financingID=new_financing.financingID
+#         )
+#         db.session.add(new_payment)
+#         db.session.commit()
+
+#         new_purchase = Purchases(
+#             bidID=bid.bidID,
+#             VIN_carID=vehicle.VIN_carID,
+#             memberID=member_id,
+#             confirmationNumber=confirmation_number_generation(),  # You may generate a confirmation number here
+#             # signature=signature_val
+#         )
+#         db.session.add(new_purchase)
+#         db.session.commit()
+
+#         # no need to modify Bid Status since its already been confirmed.
+#         return jsonify({'message': 'Vehicle purchase with financing processed successfully.'}), 200
+#     except Exception as e:
+#         db.session.rollback()
+#         return jsonify({'message': f'Error: {str(e)}'}), 500
+
+
+
+# @app.route('/api/vehicle-purchase/new-vehicle-no-finance', methods=['POST'])
+# # POST request is used by the customer to purchase a new Vehicle at BID or MSRP with NO FINANCING
+# def purchase_vehicle():
+#     member_id = session.get('member_session_id')
+#     if member_id is None:
+#         return jsonify({'message': 'You need to log in to purchase a vehicle.'}), 403
+
+#     data = request.json
+#     vehicle_vin = data.get('vehicle_vin')
+#     if vehicle_vin is None:
+#         return jsonify({'message': "Vehicle VIN is missing."}), 400
+
+#     # card_number = None
+#     # cvv = None
+#     # expiration_date = None
+
+#     payment_method = data.get('payment_method')
+#     payment_amount = data.get('payment_amount')
+#     payment_option = "Check"
+
+#     try:
+#         vehicle_cost = return_vehicle_cost(vehicle_vin)
+#     except ValueError as ve:
+#         return jsonify({'error': str(ve)}), 404  # Vehicle not found
+    
+#     if payment_method == 'MSRP':
+#         routing_number = data.get('routing_number')
+#         account_number = data.get('account_number')
+
+#         if check_ssn() is False:
+#             return jsonify({'message': 'SSN number is in Invalid Format.'}), 401
+
+#         if regex_bank_acct_check(routing_number, account_number) is False:
+#             return jsonify({'message': 'Routing Number and/or Account Number are in Invalid Formats.'}), 401
+
+#         try:
+#             return msrp_vehicle_purchase_no_financing(vehicle_vin, payment_amount, member_id, payment_option, vehicle_cost, routing_number, account_number)
+#         except Exception as e:
+#             return jsonify({'error': f'Error processing MSRP payment: {str(e)}'}), 500
+
+#     else:
+#         bid_value = data.get('bidValue')
+#         bid_status = 'Processing'
+#         try:
+#             return bid_insert_no_financing(member_id, bid_value, bid_status)
+#         except Exception as e:
+#             return jsonify({'error': f'Error processing bid: {str(e)}'}), 500
+
+
+# def msrp_vehicle_purchase_no_financing(vehicle_vin, payment_amount, member_id, payment_option, vehicle_cost, routing_number, account_number):
+#     try:
+#         # Validate signature
+#         signature_val = get_signature()
+#         if signature_val not in ['Yes', 'No']:
+#             return jsonify({'error': 'Invalid signature value.'}), 400
+#         elif signature_val == 'No':
+#             return jsonify({'error': 'You must provide a signature to purchase a vehicle.'}), 401
+
+#         # Calculate payment values
+#         valuePaid_value = payment_amount
+#         valueToPay_value = vehicle_cost - payment_amount
+
+#         # Add payment record
+#         new_payment = Payments(
+#             paymentStatus='Confirmed',
+#             valuePaid=valuePaid_value,
+#             valueToPay=valueToPay_value,
+#             initialPurchase=datetime.now(),
+#             lastPayment=datetime.now(),
+#             paymentType=payment_option,
+#             servicePurchased='Vehicle Purchase',
+#             # cardNumber=card_number,
+#             # expirationDate=expiration_date,
+#             # CVV=cvv,
+#             routingNumber=routing_number,
+#             bankAcctNumber=account_number,
+#             memberID=member_id,
+#             financingID=None
+#         )
+#         db.session.add(new_payment)
+
+#         # Commit payment record
+#         db.session.commit()
+
+#         # Add purchase record
+#         new_purchase = Purchases(
+#             VIN_carID=vehicle_vin,
+#             memberID=member_id,
+#             confirmationNumber=confirmation_number_generation(),
+#             # purchaseType='Vehicle/Add-on Purchase',
+#             # purchaseDate=datetime.now(),
+#             # signature=signature_val
+#         )
+#         db.session.add(new_purchase)
+
+#         # Commit purchase record
+#         db.session.commit()
+
+#         return jsonify({'message': 'Vehicle purchase processed successfully.'}), 200
+
+#     except IntegrityError:
+#         db.session.rollback()
+#         return jsonify({'error': 'Integrity error occurred. Please check your input data.'}), 400
+
+#     except Exception as e:
+#         db.session.rollback()
+#         return jsonify({'error': f'Error processing vehicle purchase: {str(e)}'}), 500
+
+# @app.route('/api/vehicle-purchase/new-vehicle-no-finance/bid-accepted', methods=['POST'])
+# # this endpoint is used for the customer to Purchase their cars AFTER their bid had been accepted
+# # here we only deal with the Purchases and Payments Table
+# def vehicle_purchase_bid_accepted():
+#     try:
+#         # make sure the member is logged in and has an account to purchase a vehicle and it logged in
+#         member_session_id = session.get('member_session_id')
+#         if member_session_id is None:
+#             return jsonify({'message': 'Invalid session'}), 400
+
+#         # get information sent from the frontend to here
+#         # Frontend needs to send: bid_id, payment_option, member_id, payment_amount
+#         data = request.json
+#         bid_id = data.get('bid_id')
+#         payment_option = data.get('payment_option')  # 'Card' or 'Check'
+#         member_id = data.get('member_id')
+#         payment_amount = data.get('payment_amount')
+
+#         # returns the row in BID table that matches with our bid_id
+#         bid = Bids.query.get(bid_id)
+#         if not bid or bid.bidStatus != 'Confirmed':
+#             return jsonify({'message': 'Bid not found or is Not Confirmed, Cannot continue with Purchase'}), 404
+
+#         # returns vehicle information for purchase for the vehicle to be bought
+#         vehicle = CarInfo.query.filter_by(VIN_carID=bid.VIN_carID).first()
+#         if not vehicle:
+#             return jsonify({'message': 'Vehicle not found for this bid'}), 404
+
+#         # Validate and retrieve card or check information based on payment_option
+#         if payment_option == 'Card':
+#             card_number = data.get('card_number')
+#             cvv = data.get('CVV')
+#             expiration_date = data.get('expirationDate')
+#             routing_number = None
+#             account_number = None
+#             regex_card_check(card_number, cvv, expiration_date)
+#             if payment_amount > 5000:  # cards cannot pay more than 5000
+#                 return jsonify({'message': 'Choose A lower value for Bank Cards'}), 400
+#         elif payment_option == 'Check':
+#             routing_number = data.get('routing_number')
+#             account_number = data.get('account_number')
+#             card_number = None
+#             cvv = None
+#             expiration_date = None
+#             regex_bank_acct_check(routing_number, account_number)
+#         else:
+#             return jsonify({'message': 'Invalid payment option.'}), 400
+
+#         # Retrieve vehicle cost
+#         # vehicle_cost = return_vehicle_cost(vehicle_vin) # no need because the cost is based on the confirmed bid
+#         total_valuePaid = bid.bidValue - payment_amount
+#         # signature = get_signature()
+
+#         # create a payment entry to insert into the DB for the new payment/purchase information
+#         new_payment = Payments(
+#             paymentStatus='Confirmed',
+#             valuePaid=payment_amount,
+#             valueToPay=total_valuePaid,
+#             initialPurchase=datetime.now(),
+#             lastPayment=datetime.now(),
+#             paymentType=payment_option,
+#             cardNumber=card_number,
+#             expirationDate=expiration_date,
+#             CVV=cvv,
+#             routingNumber=routing_number,
+#             bankAcctNumber=account_number,
+#             memberID=member_id
+#         )
+#         db.session.add(new_payment)
+#         db.session.commit()
+
+#         new_purchase = Purchases(
+#             VIN_carID=vehicle.VIN_carID,
+#             memberID=member_id,
+#             confirmationNumber=confirmation_number_generation(),  # You may generate a confirmation number here
+#             # signature=signature_val
+#         )
+#         db.session.add(new_purchase)
+#         db.session.commit()
+#         return jsonify({'message': 'Vehicle purchase processed successfully.'}), 200
+#     except Exception as e:
+#         return jsonify({'error': 'Bid not found for the specified member and vehicle, could not purchase vehicle'}), 404
+
